Always return after the error is sent back
This ensures that an extra call does not get made to the server.
